# src/main/resources/META-INF/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
...
    DienstRequest:
      description: Request object voor het toevoegen van een dienst aan een dienstverlener
      type: object
      required:
      - naam
      properties:
        naam:
          type: string
          pattern: ".*\\S.*"
        beschrijving:
          type: string
    DienstResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        naam:
          type: string
        beschrijving:
          type: string
    DienstverlenerRequest:
      description: Request object voor het toevoegen van een dienstverlener
      type: object
      required:
      - naam
      properties:
        naam:
          type: string
        beschrijving:
          type: string
    DienstverlenerResponse:
      type: object
      properties:
        naam:
          type: string
        beschrijving:
          type: string
        diensten:
          type: array
          items:
            $ref: "#/components/schemas/DienstResponse"
    EmailVerificatieCodeAanvraagRequest:
      type: object
      required:
      - email
      - identificatieNummer
      - identificatieType
      properties:
        email:
          type: string
          pattern: ".*\\S.*"
        identificatieNummer:
          type: string
          pattern: ".*\\S.*"
        identificatieType:
          $ref: "#/components/schemas/IdentificatieType"
    EmailVerificatieRequest:
      description: Request object voor het verifiëren van een emailadres
      type: object
      required:
      - email
      - identificatieNummer
      - identificatieType
      - verificatieCode
      properties:
        email:
          type: string
        identificatieNummer:
          type: string
        identificatieType:
          $ref: "#/components/schemas/IdentificatieType"
        verificatieCode:
          type: string
    HttpProblem:
      description: HTTP Problem Response according to RFC9457 and RFC7807
      additionalProperties: true
      type: object
      properties:
        type:
          type: string
          format: uri
          description: A optional URI reference that identifies the problem type
          examples:
          - https://example.com/errors/not-found
        title:
          type: string
          description: "A optional, short, human-readable summary of the problem type"
          examples:
          - Not Found
        status:
          type: integer
          format: int32
          description: The HTTP status code for this occurrence of the problem
          examples:
          - 404
        detail:
          type: string
          description: A optional human-readable explanation specific to this occurrence
            of the problem
          examples:
          - Record not found
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence of
            the problem
          examples:
          - https://api.example.com/errors/123
...
    IdentificatieType:
      type: string
      enum:
      - BSN
      - KVK
      - RSIN
...
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
...
paths:
...
  /api/profielservice/v1/dienstverlener:
    post:
      summary: Voegt een dienstverlener toe
      operationId: addDienstverlener
      description: Voegt een nieuwe dienstverlener toe met optionele beschrijving.
      tags:
      - Dienstverlener
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DienstverlenerRequest"
        required: true
      responses:
        "500":
          description: Interne serverfout
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "201":
          description: Dienstverlener succesvol toegevoegd
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DienstverlenerResponse"
        "400":
          description: Request body ontbreekt of is ongeldig
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "409":
          description: Dienstverlener bestaat al met conflicterende waarden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
  /api/profielservice/v1/dienstverlener/{dienstverlenerNaam}/diensten:
    post:
      summary: Voegt een dienst toe aan een dienstverlener
      operationId: addDienstToDienstverlener
      description: Voegt een nieuwe dienst toe met beschrijving aan een bestaande
        dienstverlener.
      tags:
      - Dienstverlener
      parameters:
      - name: dienstverlenerNaam
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DienstRequest"
        required: true
      responses:
        "500":
          description: Interne serverfout
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "201":
          description: Dienst succesvol toegevoegd
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DienstResponse"
        "400":
          description: Request body ontbreekt of is ongeldig
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "409":
          description: Dienst bestaat al met een andere beschrijving
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
  /api/profielservice/v1/dienstverlener/{naam}:
    get:
      summary: Vraagt gegevens van dienstverlener
      operationId: getDienstverlener
      description: "Geeft gegevens van gevraagde dienstverlener terug, inclusief de\
        \ aangesloten diensten."
      tags:
      - Dienstverlener
      parameters:
      - name: naam
        in: path
        required: true
        schema:
          type: string
      responses:
        "500":
          description: Interne serverfout
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "200":
          description: Dienstverlener succesvol opgehaald
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DienstverlenerResponse"
        "404":
          description: Dienstverlener niet gevonden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
  /api/profielservice/v1/emailverificatie:
    post:
      summary: Verifieer een email met een verificatie code
      operationId: verifieerEmail
      description: Verifieert een email adres aan de hand van de eerder aangevraagde
        verificatie code.
      tags:
      - EmailVerificatie
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailVerificatieRequest"
        required: true
      responses:
        "500":
          description: Interne serverfout
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "200":
          description: Email verificatie succesvol
          content:
            application/json:
              schema: {}
        "400":
          description: Email verificatie mislukt
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
  /api/profielservice/v1/emailverificatie/code:
    post:
      summary: (Opnieuw) aanvragen voor een code van een (al geverifieerde) mail adres
      operationId: vraagEmailVerificatieCodeAan
      description: "Vraagt een email verificatie code aan. Let op, bij het aanmaken\
        \ van een profiel wordt al een email verificatie code aangevraagd. Dit is\
        \ voor het opnieuw aanvragen van een code."
      tags:
      - EmailVerificatie
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailVerificatieCodeAanvraagRequest"
        required: true
      responses:
        "500":
          description: Interne serverfout
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "200":
          description: Email verificatie code aanvraag succesvol
          content:
            application/json:
              schema: {}
        "400":
          description: Request body ontbreekt of is ongeldig
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "404":
          description: Partij of Contactgegeven niet gevonden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "503":
          description: Service tijdelijk niet beschikbaar
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
...
info:
  title: profiel-service API
  version: 1.0.0-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value